Add official riscv-tests suite + fix three pipeline bugs
Integrates the official RISC-V ISA test suite (rv32ui-p-* and rv32um-p-*,
49 ELFs) as both correctness tests and performance workloads across all three
pipeline configurations (SingleCycle, FiveStage, OoOE). All 651 tests pass.

Test infrastructure:
- riscv-tests submodule with custom env/ (EBREAK halt, gp pass/fail
  convention, link at 0x0 for FlatMemory compatibility)
- Makefile builds 41 rv32ui + 8 rv32um ELFs; excludes fence_i.S
  (requires zifencei) and ma_data.S (trap-handler test incompatible
  with FlatMemory's native misaligned access)
- RiscVTestSuiteTests: Theory tests per pipeline + PerfSummary_AllTests
  reporting fixture

Bug fixes found during integration:
- sltiu (RvExecutor): 64-bit sign-extension caused wrong comparison;
  fixed to (uint)rs1 < (uint)imm
- HazardUnit.Forward (FiveStage): EX/MEM forwarding was applied before
  MEM/WB, so MEM/WB overwrote the more-recent EX/MEM result when two
  consecutive instructions wrote the same register; fixed by applying
  MEM/WB first so EX/MEM wins
- HasPrecedingPendingStore (OoOE): loads were allowed to issue once a
  preceding store executed, but stores write to memory only at commit;
  fixed to block loads until all preceding stores leave the ROB

+++ RiscV/Execute/RvExecutor.cs
@@ -59,7 +59,7 @@ public sealed class RvExecutor : IExecutor {
             RvSlti (_, var rs1, var imm) =>
                 Reg((int)regs.Read(rs1) < imm ? 1UL : 0UL),
             RvSltiu(_, var rs1, var imm) =>
-                Reg(regs.Read(rs1) < (ulong)imm ? 1UL : 0UL),
+                Reg((uint)regs.Read(rs1) < (uint)imm ? 1UL : 0UL),
 
             // ── Loads ─────────────────────────────────────────────────────────
             RvLb (_, var rs1, var imm) => Load(
